что-то более менее рабочее WIP

+++ src/vars.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-extern float r_x, r_y, r_w, r_h; //координата x, координата y, ширина, высота 
+extern int r_x, r_y, r_w, r_h; //координата x, координата y, ширина, высота
